TIP-841: Move the channel controllers to Akeneo\Channel
