+++ src/Akeneo/Channel/Bundle/Controller/ExternalApi/ChannelController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\Controller;
+namespace Akeneo\Channel\Bundle\Controller\ExternalApi;
 
 use Akeneo\Tool\Bundle\ApiBundle\Documentation;
 use Akeneo\Tool\Bundle\ApiBundle\Stream\StreamResourceResponse;
