+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/controllers.yml
@@ -5,7 +5,6 @@ parameters:
     pim_api.controller.attribute.class: Akeneo\Tool\Bundle\ApiBundle\Controller\AttributeController
     pim_api.controller.attribute_option.class: Akeneo\Tool\Bundle\ApiBundle\Controller\AttributeOptionController
     pim_api.controller.attribute_group.class: Akeneo\Tool\Bundle\ApiBundle\Controller\AttributeGroupController
-    pim_api.controller.channel.class: Akeneo\Tool\Bundle\ApiBundle\Controller\ChannelController
     pim_api.controller.token.class: Akeneo\Tool\Bundle\ApiBundle\Controller\TokenController
     pim_api.controller.product.class: Akeneo\Tool\Bundle\ApiBundle\Controller\ProductController
     pim_api.controller.product_model.class: Akeneo\Tool\Bundle\ApiBundle\Controller\ProductModelController
@@ -107,21 +106,6 @@ services:
             - '@router'
             - '@pim_catalog.saver.attribute_group'
             - '@pim_api.stream.attribute_group_partial_update_stream'
-            - '%pim_api.configuration%'
-
-    pim_api.controller.channel:
-        class: '%pim_api.controller.channel.class%'
-        arguments:
-            - '@pim_api.repository.channel'
-            - '@pim_serializer'
-            - '@pim_api.pagination.offset_hal_paginator'
-            - '@pim_api.pagination.parameter_validator'
-            - '@pim_catalog.factory.channel'
-            - '@pim_catalog.updater.channel'
-            - '@validator'
-            - '@router'
-            - '@pim_catalog.saver.channel'
-            - '@pim_api.stream.channel_partial_update_stream'
             - '%pim_api.configuration%'
 
     pim_api.controller.token:
